requirements.txt
absl-py==1.4.0
annotated-types==0.7.0
astunparse==1.6.3
bcrypt==4.3.0
cachetools==5.3.1
# Editable install with no version control (car_rental==0.1)
-e d:\car_rental_system-main
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.1.0
cryptography==44.0.2
cvlib==0.2.7
cycler==0.11.0
dill==0.3.9
exceptiongroup==1.2.2
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.20.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.54.2
h5py==3.8.0
httplib2==0.22.0
ibm-cloud-sdk-core==3.23.0
ibm-platform-services==0.63.0
idna==3.4
imageio==2.31.3
imutils==0.5.4
iniconfig==2.1.0
jax==0.4.12
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
mpmath==1.3.0
mysql-connector-python==9.3.0
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.2.3
pbr==6.1.1
Pillow==9.5.0
pluggy==1.5.0
prettytable==3.16.0
progressbar==2.5
proto-plus==1.23.0
protobuf==4.23.3
pyarrow==20.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.4
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.0.9
pyspnego==0.11.2
pytest==8.3.5
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
qiskit==2.0.0
qiskit-ibm-runtime==0.37.0
requests==2.31.0
requests-oauthlib==1.3.1
requests_ntlm==1.3.0
rsa==4.9
rustworkx==0.16.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sspilib==0.3.0
stevedore==5.4.1
symengine==0.13.0
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tomli==2.2.1
typing_extensions==4.6.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.3.6
wrapt==1.14.1
